refactor(project): split extraction by domain
