+++ crates/djls-project/src/lib.rs
@@ -1,6 +1,5 @@
 mod ast;
 mod db;
-mod extraction;
 mod names;
 mod parse;
 mod project;
@@ -18,19 +17,19 @@ pub use db::Db;
 pub use templates::InactiveLibraries;
 pub use templates::InactiveLibrary;
 pub use templates::inactive_template_libraries;
-pub use extraction::DjangoSettings;
-pub use extraction::InstalledAppsSetting;
-pub use extraction::RegistrationInfo;
-pub use extraction::RegistrationKind;
-pub use extraction::SettingsSource;
-pub use extraction::SettingsSourceResolver;
-pub use extraction::SettingsStarImport;
-pub use extraction::StaticKnowledge;
-pub use extraction::TemplateBackend;
-pub use extraction::TemplateDirPath;
-pub use extraction::TemplateSettings;
-pub use extraction::collect_registrations_from_body;
-pub use extraction::extract_settings;
+pub use settings::DjangoSettings;
+pub use settings::InstalledAppsSetting;
+pub use settings::SettingsSource;
+pub use settings::SettingsSourceResolver;
+pub use settings::SettingsStarImport;
+pub use settings::StaticKnowledge;
+pub use settings::TemplateBackend;
+pub use settings::TemplateDirPath;
+pub use settings::TemplateSettings;
+pub use settings::extract_settings;
+pub use templates::RegistrationInfo;
+pub use templates::RegistrationKind;
+pub use templates::collect_registrations_from_body;
 pub use names::LibraryName;
 pub use names::ModulePath;
 pub use names::PyModuleName;
