+++ crates/djls-project/src/settings/extraction.rs
@@ -15,18 +15,18 @@ use rustc_hash::FxHashMap;
 use rustc_hash::FxHashSet;
 
 use crate::ast::ExprExt;
-use crate::extraction::paths::evaluate_path_expr;
-use crate::extraction::settings::DjangoSettings;
-use crate::extraction::settings::InstalledAppsSetting;
-use crate::extraction::settings::LocalBindings;
-use crate::extraction::settings::Reason;
-use crate::extraction::settings::SettingsSource;
-use crate::extraction::settings::SettingsSourceResolver;
-use crate::extraction::settings::SettingsStarImport;
-use crate::extraction::settings::StaticKnowledge;
-use crate::extraction::settings::TemplateBackend;
-use crate::extraction::settings::TemplateDirPath;
-use crate::extraction::settings::TemplateSettings;
+use crate::settings::paths::evaluate_path_expr;
+use crate::settings::types::DjangoSettings;
+use crate::settings::types::InstalledAppsSetting;
+use crate::settings::types::LocalBindings;
+use crate::settings::types::Reason;
+use crate::settings::types::SettingsSource;
+use crate::settings::types::SettingsSourceResolver;
+use crate::settings::types::SettingsStarImport;
+use crate::settings::types::StaticKnowledge;
+use crate::settings::types::TemplateBackend;
+use crate::settings::types::TemplateDirPath;
+use crate::settings::types::TemplateSettings;
 
 const INSTALLED_APPS: &str = "INSTALLED_APPS";
 const TEMPLATES: &str = "TEMPLATES";
