+++ crates/djls-project/src/templates/registrations.rs
